+++ Mate.DataCore/DataModel/T_StockExchange.cs
@@ -14,11 +14,11 @@ namespace Mate.DataCore.DataModel
         public SimulationType SimulationType { get; set; }
         public int SimulationNumber { get; set; }
         public M_Stock Stock { get; set; }
-        public int RequiredOnTime { get; set; }
+        public long RequiredOnTime { get; set; }
         public State State { get; set; }
         public decimal Quantity { get; set; }
         // time of withdrawal/insert (wird von der Simulation belegt bei erfuellung des Requests)id
-        public int Time { get; set; }
+        public long Time { get; set; }
         public ExchangeType ExchangeType { get; set; }
 
         public StockExchangeType StockExchangeType { get; set; }
