+++ Mate.Production.Core/Environment/Options/SettlingStart.cs
@@ -3,9 +3,9 @@ using Mate.Production.Core.Environment.Abstractions;
 
 namespace Mate.Production.Core.Environment.Options
 {
-    public class SettlingStart : Option<int>
+    public class SettlingStart : Option<long>
     {
-        public SettlingStart(int value)
+        public SettlingStart(long value)
         {
             _value = value;
         }
@@ -13,7 +13,7 @@ namespace Mate.Production.Core.Environment.Options
         public SettlingStart()
         {
             Action = (config, argument) => {
-                config.AddOption(o: new SettlingStart(value: int.Parse(s: argument)));
+                config.AddOption(o: new SettlingStart(value: long.Parse(s: argument)));
             };
         }
     }
